requirements.txt
torch-cluster==1.5.8
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0
ogb>=1.3.0
scikit-learn>=0.23.2
dgl
tensorboard
tqdm
simplicialX